Pokemon entity fixed

+++ libs/backend/features/src/lib/pokemon/pokemon.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { IPokemon } from '@pokemon/shared/api';
 import { CreatePokemonDto, UpdatePokemonDto } from '@pokemon/backend/dto';
 import { Logger } from '@nestjs/common';
@@ -7,6 +7,7 @@ import { Model } from 'mongoose';
 import { Pokemon as PokemonModel, PokemonDocument } from './pokemon.schema';
 import { Team as TeamModel, TeamDocument } from '../team/team.schema';
 import { TokenBlacklistService } from '../user/blacklist.service';
+import { User as UserModel, UserDocument } from '../user/user.schema';
 
 @Injectable()
 export class PokemonService {
@@ -16,6 +17,7 @@ export class PokemonService {
   constructor(
         @InjectModel(PokemonModel.name) private pokemonModel: Model<PokemonDocument>,
         @InjectModel(TeamModel.name) private teamModel: Model<TeamDocument>,
+        @InjectModel(UserModel.name) private userModel: Model<UserDocument>,
         private readonly tokenBlackListService: TokenBlacklistService
         ) {}
 
@@ -34,20 +36,93 @@ export class PokemonService {
         return item;
     }
 
-    async create(pokemon: CreatePokemonDto): Promise<IPokemon> {
+    async create(pokemon: CreatePokemonDto, userId: string): Promise<IPokemon> {
         this.logger.log(`create(${pokemon})`);
+        const user = await this.userModel.findOne({userId: userId}).lean().exec();
+        this.logger.log(`userId: ` + userId, `User: ` + user);
+
+        if(!user){
+            throw new HttpException(
+                {
+                    status: HttpStatus.UNAUTHORIZED,
+                    error: 'Unauthorized',
+                    message: 'User not found',
+                },
+                HttpStatus.UNAUTHORIZED
+            );
+        }
+
         const lowestAvailableId = await this.getLowestAvailablePokemonId();
-        const pokemonWithId = { ...pokemon, pokemonId: lowestAvailableId };
+        const pokemonWithId = { ...pokemon, pokemonId: lowestAvailableId, creator: Number(userId) }; 
+        pokemon.creator = user.userId;
         const createdItem = await this.pokemonModel.create(pokemonWithId);
         return createdItem;
     }
 
-    async update(pokemonId: number, pokemon: UpdatePokemonDto): Promise<IPokemon | null> {
+    async update(userId: string, pokemonId: number, pokemon: UpdatePokemonDto): Promise<IPokemon | null> {
+        this.logger.log(`Update pokémon ${pokemon.name}`);
+
+        const existingPokemon = await this.pokemonModel.findOne({pokemonId}).lean().exec();
+        const user = await this.userModel.findOne({userId}).lean().exec();
+
+        if(!existingPokemon || !user){
+            throw new HttpException(
+                { status: HttpStatus.NOT_FOUND, error: 'Not Found', message: 'Pokémon or user not found' },
+                HttpStatus.NOT_FOUND
+            );
+        }
+
+        if(user.role !== 'Admin' && user.userId !== existingPokemon.creator){
+            throw new HttpException(
+                { status: HttpStatus.UNAUTHORIZED, error: 'Unauthorized', message: 'You do not have permission to update this pokémon' },
+                HttpStatus.UNAUTHORIZED
+            );
+        }
+        
+        const updatedPokemon = await this.pokemonModel.findOneAndUpdate({pokemonId}, pokemon, { new: true}).exec();
+        if(!updatedPokemon) return null;
+
         return this.pokemonModel.findOneAndUpdate({ pokemonId }, pokemon);
     }
 
-    async delete(pokemonId: number): Promise<IPokemon | null> {
+    async delete(userId: string, pokemonId: number): Promise<IPokemon | null> {
+        this.logger.log(`Delete pokemon ${pokemonId}`);
+        const pokemon = await this.pokemonModel.findOne({pokemonId}).lean().exec();
+        const user = await this.userModel.findOne({userId: userId}).lean().exec();
+
+        if(!pokemon || !user){
+            this.logger.warn(`Pokémon or user not found`);
+            throw new HttpException(
+                {
+                    status: HttpStatus.NOT_FOUND,
+                    error: 'Not Found',
+                    message: `Pokémon or user not found`
+                },
+                HttpStatus.NOT_FOUND
+            );
+        }
+        if(user?.role !== 'Admin' && user?.userId !== pokemon?.creator){
+            throw new HttpException(
+                {
+                    status: HttpStatus.UNAUTHORIZED,
+                    error: 'Unauthorized',
+                    message: 'You do not have permission to delete this pokémon',
+                },
+                HttpStatus.UNAUTHORIZED
+            ); 
+        }
         const deletedPokemon = await this.pokemonModel.findOneAndDelete({ pokemonId }, {}).exec();
+        if(!deletedPokemon){
+            throw new HttpException(
+                {
+                    status: HttpStatus.NOT_FOUND,
+                    error: 'Not Found',
+                    message: `Pokémon with id ${pokemonId} not found for deletion`
+                },
+                HttpStatus.NOT_FOUND
+            );
+        }
+        this.logger.log(`Deleted pokémon with id ${pokemonId}`);
         return deletedPokemon;
     }
